% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_content.R
\name{gc_content}
\alias{gc_content}
\title{Calculate GC content of sequencee}
\usage{
gc_content(my_seq)
}
\arguments{
\item{my_seq}{character vector}
}
\value{
: vector of \%GC values (what \% of the bases are G or C)
}
\description{
this function will take in a vector of sequences and calculate the GC content
}
